requirements.txt
python-novaclient==17.6.0
keystoneauth1==4.4.0